+++ CS_HomeWork_Day3_04/StructRequest.cs
@@ -11,7 +11,7 @@ namespace CS_HomeWork_Day3_04
             public string Client { get; private set; }
             public DateTime OrderDate { get; private set; }
             public int[] OrderedItems { get; private set; } // Предпологалось что перечень товаров это ID товара
-            public double OrderSum { get { return (double)OrderedItems.Length; } }
+            public decimal OrderSum { get { return (double)OrderedItems.Length; } }
             public PayType payType { get; private set; }
         }
     }
